Dynamically determine the Images account to use
This is done by determine the account "type" (Production or Staging)
of the Shared Services account, then looking for the Images account of
the same type.

Also stop associating public IPs with the FreeIPA master and replica.
Those instances are in private networks now, so they don't need public
IPs or public DNS records.