Change submodule to allow for assigning PTR records in different zones
I realized that this module was incorrectly creating all the PTR
records for the IPA servers in the same reverse zone.  The reverse
zone ID needs to be specified in the IP map object as well as the IP,
so that each reverse record can be created in its correct zone.

This contributed to the LDAP connection errors that @dav3r and I saw
yesterday when deploying.