Explicitly set the IPs of the IPA servers
The latest version of cisagov/freeipa-server-tf-module requires this
in order to break the dependency of DNS record resources on the
corresponding EC2 instance resources; otherwise, it is not possible to
recreate the EC2 instances one by one when a new FreeIPA server AMI is
made available.

Also remove the realm from the module configuration for replicas since
it is not used there, and add the "advertise" attribute to the hosts
map as required by the DNS submodule.