Add a provider and associated remote state
The new provider is used to create an SSM Parameter Store read role so
that we can access some sensitive data items necessary to link the
Nessus Agent with the Tenable/Nessus server.

We also remove an unused provider and associated remote state.  These
were added in the event that we wanted to create public DNS records
for the IPA servers, but that is no longer supported.