Reduce default TTL from 3600 to 60
It makes sense to reduce the default TTL to 60, since I intend to
eventually add code to automatically update FreeIPA DNS records
depending on the status of health checks.  Making this change now also
means that there is one less step I have to do when I upgrade the
FreeIPA cluster, and I also don't have to wait an hour (for the
previous default TTL of 3600 to time out) before making changes.