Refactor the three IPA server modules into a single map of modules
This allows us to simplify the LB target group attachments
considerably.

Note that this does leave the state in a slightly janky state, since
Terraform now wants to replace ipa1 and ipa2.  It wants to do this
because we didn't used to specify the realm input variable for those
two servers, since replicas never make use of that input variable.
Now that we are treating all servers identically, though, we specify
the realm for all three.  This causes a change to the instance
metadata that doesn't actually change anything, but Terraform detects
the change and wants to redeploy.  This jankiness will go away the
next time the FreeIPA instances are redeployed.